refactor: centralize shared validation into portfolio-core
Business validation (config allowlist/value checks, UUID format, holding
fields, target-weight budget, dividend/alert/account fields) was duplicated
between src-tauri/src/commands/ and portfolio-mcp/src/validation.rs, risking
drift when a rule changed in only one place.

Move the transport-neutral rules into portfolio-core/src/validation/ as
plain Result<T, String> functions with their own test suite. Both the Tauri
commands and portfolio-mcp/src/validation.rs now delegate to these shared
functions and only adapt the result to their own error type (AppError vs
McpError), preserving every existing user-facing error message.

Closes #758
