+++ portfolio-core/Cargo.toml
@@ -5,9 +5,11 @@ edition = "2021"
 
 [dependencies]
 serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 chrono = { version = "0.4", features = ["serde"] }
 tracing = "0.1"
 ts-rs = { version = "12", features = ["serde-compat"], optional = true }
+uuid = "1"
 
 [features]
 ts = ["dep:ts-rs"]
