+++ portfolio-core/src/lib.rs
@@ -2,3 +2,4 @@ pub mod fx;
 pub mod snapshot;
 pub mod stress;
 pub mod types;
+pub mod validation;
